Thinking of logic for multi-type sample 2d slice plots.

+++ year2/parametric_tests/error_tests/error_test.py
@@ -7,7 +7,7 @@ import argparse
 
 def tester(g1_p1, g2_g1, g3_g2, g4_g3, core):
 
-    Dir = "core{}/".format(int(core))
+    Dir = "core{}/".format(core)
     np.savetxt("files/runs/{}placeholder.txt".format(Dir),[0])
     try:
         eos = lalsim.SimNeutronStarEOS4ParameterSpectralDecomposition(g1_p1, g2_g1, g3_g2, g4_g3)
@@ -22,7 +22,7 @@ if __name__ == '__main__':
     parser.add_argument("g2_g1", help="Second Parameter", type=float)
     parser.add_argument("g3_g2", help="Third Parameter", type=float)
     parser.add_argument("g4_g3", help="Fourth Parameter", type=float)
-    parser.add_argument("core", help="Core Number", type=float)
+    parser.add_argument("core", help="Core Number", type=int)
     args = parser.parse_args()
 
     tester(args.g1_p1,args.g2_g1,args.g3_g2,args.g4_g3,args.core)
